REL: rename chxanalys -> pyCHX
