+++ pyCHX/Badpixels.py
