+++ pyCHX/chx_compress_analysis.py
@@ -5,13 +5,13 @@ import struct
 
 import matplotlib.pyplot as plt
 
-from chxanalys.chx_libs import (np, roi, time, datetime, os,  getpass, db, get_images,LogNorm,Figure, RUN_GUI)
-#from chxanalys.chx_generic_functions import (get_circular_average)
-#from chxanalys.XPCS_SAXS import (get_circular_average)
-from chxanalys.chx_libs import  ( colors, markers, colors_,  markers_)
+from pyCHX.chx_libs import (np, roi, time, datetime, os,  getpass, db, get_images,LogNorm,Figure, RUN_GUI)
+#from pyCHX.chx_generic_functions import (get_circular_average)
+#from pyCHX.XPCS_SAXS import (get_circular_average)
+from pyCHX.chx_libs import  ( colors, markers, colors_,  markers_)
 
 import os
-from chxanalys.chx_generic_functions import ( save_arrays )
+from pyCHX.chx_generic_functions import ( save_arrays )
 
 
 from skbeam.core.utils import multi_tau_lags
@@ -21,11 +21,11 @@ from collections import namedtuple
 import logging
 logger = logging.getLogger(__name__)
 
-from chxanalys.chx_compress import   (compress_eigerdata, read_compressed_eigerdata,init_compress_eigerdata, 
+from pyCHX.chx_compress import   (compress_eigerdata, read_compressed_eigerdata,init_compress_eigerdata, 
                                       Multifile,pass_FD,get_avg_imgc,mean_intensityc, get_each_frame_intensityc)
 
 from modest_image import imshow                                      
-#from chxanalys.chx_compress import *
+#from pyCHX.chx_compress import *
 
 
 def get_time_edge_avg_img(FD, frame_edge,show_progress =True):
