+++ pyCHX/chx_correlationp.py
@@ -8,14 +8,14 @@ from __future__ import absolute_import, division, print_function
 from skbeam.core.utils import multi_tau_lags
 from skbeam.core.roi import extract_label_indices
 
-from chxanalys.chx_libs import tqdm
-from chxanalys.chx_correlationc import (  get_pixelist_interp_iq, _validate_and_transform_inputs,
+from pyCHX.chx_libs import tqdm
+from pyCHX.chx_correlationc import (  get_pixelist_interp_iq, _validate_and_transform_inputs,
                  _one_time_process as _one_time_processp,   
                 _two_time_process  as _two_time_processp )
 
 
 
-from chxanalys.chx_compress import ( run_dill_encoded,apply_async, map_async,pass_FD, go_through_FD  ) 
+from pyCHX.chx_compress import ( run_dill_encoded,apply_async, map_async,pass_FD, go_through_FD  ) 
 
 
 from multiprocessing import Pool
