+++ pyCHX/chx_packages.py
@@ -2,16 +2,16 @@ import pickle as cpk
 from skimage.draw import line_aa, line, polygon
 
 
-from chxanalys.chx_libs import (np, roi, time, datetime, os, get_events, 
+from pyCHX.chx_libs import (np, roi, time, datetime, os, get_events, 
                             getpass, db, get_images,LogNorm, plt,tqdm, utils, Model,
                            multi_tau_lags,  random,  warnings)
-from chxanalys.chx_libs import (cmap_vge, cmap_albula, Javascript, EigerHandler, EigerHandler, pims,
+from pyCHX.chx_libs import (cmap_vge, cmap_albula, Javascript, EigerHandler, EigerHandler, pims,
                                 h5py)
 
 
 
 
-from chxanalys.chx_generic_functions import (get_detector, get_fields, get_sid_filenames,  
+from pyCHX.chx_generic_functions import (get_detector, get_fields, get_sid_filenames,  
  load_data, load_mask,get_fields, reverse_updown, ring_edges,get_avg_img,check_shutter_open,
 apply_mask, show_img,check_ROI_intensity,run_time, plot1D, get_each_frame_intensity, 
 create_hot_pixel_mask,show_ROI_on_image,create_time_slice,save_lists, 
@@ -37,7 +37,7 @@ create_hot_pixel_mask,show_ROI_on_image,create_time_slice,save_lists,
                                             )
  
 
-from chxanalys.XPCS_SAXS import (get_circular_average,save_lists,get_ring_mask, get_each_ring_mean_intensity,
+from pyCHX.XPCS_SAXS import (get_circular_average,save_lists,get_ring_mask, get_each_ring_mean_intensity,
                              plot_qIq_with_ROI, cal_g2, create_hot_pixel_mask,get_circular_average,get_t_iq, 
                               get_t_iqc,multi_uids_saxs_xpcs_analysis,
                               plot_t_iqc,  plot_circular_average, get_seg_from_ring_mask, 
@@ -47,40 +47,40 @@ from chxanalys.XPCS_SAXS import (get_circular_average,save_lists,get_ring_mask, 
                                 )
 
 
-from chxanalys.Two_Time_Correlation_Function import (show_C12, get_one_time_from_two_time, 
+from pyCHX.Two_Time_Correlation_Function import (show_C12, get_one_time_from_two_time, 
                                             get_four_time_from_two_time,rotate_g12q_to_rectangle)
-from chxanalys.chx_compress import (combine_binary_files,
+from pyCHX.chx_compress import (combine_binary_files,
                        segment_compress_eigerdata,     create_compress_header,            
                         para_segment_compress_eigerdata,para_compress_eigerdata)
 
-from chxanalys.chx_compress_analysis import ( compress_eigerdata, read_compressed_eigerdata,
+from pyCHX.chx_compress_analysis import ( compress_eigerdata, read_compressed_eigerdata,
                                          Multifile,get_avg_imgc, get_each_frame_intensityc,
             get_each_ring_mean_intensityc, mean_intensityc,cal_waterfallc,plot_waterfallc, 
                     cal_each_ring_mean_intensityc,plot_each_ring_mean_intensityc, get_time_edge_avg_img
 )
 
-from chxanalys.SAXS import fit_form_factor,show_saxs_qmap
-from chxanalys.chx_correlationc import ( cal_g2c,Get_Pixel_Arrayc,auto_two_Arrayc,get_pixelist_interp_iq,)
-from chxanalys.chx_correlationp import (cal_g2p, auto_two_Arrayp)
+from pyCHX.SAXS import fit_form_factor,show_saxs_qmap
+from pyCHX.chx_correlationc import ( cal_g2c,Get_Pixel_Arrayc,auto_two_Arrayc,get_pixelist_interp_iq,)
+from pyCHX.chx_correlationp import (cal_g2p, auto_two_Arrayp)
 
-from chxanalys.Create_Report import (create_pdf_report, 
+from pyCHX.Create_Report import (create_pdf_report, 
                             create_multi_pdf_reports_for_uids,create_one_pdf_reports_for_uids,
                                     make_pdf_report, export_xpcs_results_to_h5, extract_xpcs_results_from_h5 )
 
-from chxanalys.chx_olog import (LogEntry,Attachment, update_olog_uid, update_olog_id,
+from pyCHX.chx_olog import (LogEntry,Attachment, update_olog_uid, update_olog_id,
                                 update_olog_uid_with_file)
 
-from chxanalys.XPCS_GiSAXS import (get_qedge,get_qmap_label,get_qr_tick_label, get_reflected_angles,
+from pyCHX.XPCS_GiSAXS import (get_qedge,get_qmap_label,get_qr_tick_label, get_reflected_angles,
 convert_gisaxs_pixel_to_q, show_qzr_map, get_1d_qr, get_qzrmap, show_qzr_roi,get_each_box_mean_intensity,
  plot_gisaxs_two_g2,plot_qr_1d_with_ROI,fit_qr_qz_rate,
         multi_uids_gisaxs_xpcs_analysis,plot_gisaxs_g4,get_t_qrc, plot_t_qrc,
                                   get_qzr_map, plot_qzr_map, get_gisaxs_roi, cal_1d_qr, 
                                   get_t_qrc,  plot_qrt_pds )
 
-from chxanalys.chx_specklecp import  ( xsvsc, xsvsp, get_xsvs_fit,plot_xsvs_fit, save_KM,plot_g2_contrast,
+from pyCHX.chx_specklecp import  ( xsvsc, xsvsp, get_xsvs_fit,plot_xsvs_fit, save_KM,plot_g2_contrast,
                                       get_binned_his_std, get_contrast, save_bin_his_std, get_his_std_from_pds )
 
 
-from chxanalys.DataGonio import (qphiavg)
+from pyCHX.DataGonio import (qphiavg)
