+++ pyCHX/chx_specklecp.py
@@ -33,8 +33,8 @@ from multiprocessing import Pool
 import dill
 import itertools
 
-from chxanalys.chx_compress import ( run_dill_encoded,apply_async, map_async,pass_FD, go_through_FD  ) 
-from chxanalys.chx_generic_functions import trans_data_to_pd
+from pyCHX.chx_compress import ( run_dill_encoded,apply_async, map_async,pass_FD, go_through_FD  ) 
+from pyCHX.chx_generic_functions import trans_data_to_pd
 
 
 def xsvsp(FD, label_array,  only_two_levels= True,
