+++ pyCHX/xpcs_timepixel_debug.py
@@ -5,8 +5,7 @@ from numpy import around,intersect1d, ravel, unique,hstack,vstack,zeros_like
 from numpy import save, load, dot
 from numpy.linalg import lstsq
 from numpy import polyfit,poly1d;
-import sys,os
-import pickle as pkl
+import sys
 
 import matplotlib.pyplot as plt 
 #from Init_for_Timepix import * # the setup file 
@@ -16,180 +15,19 @@ import struct
 import numpy as np    
 from tqdm import tqdm    
 import pandas as pds
-from chxanalys.chx_libs import multi_tau_lags
-from chxanalys.chx_compress import Multifile,  go_through_FD, pass_FD
-
-def get_FD_end_num(FD, maxend=1e10):
-    N = maxend
-    for i in range(0,int(maxend)):
-        try:
-            FD.seekimg(i)
-        except:
-            N = i 
-            break
-    FD.seekimg(0)
-    return N
-
-def compress_timepix_data( pos, t, tbins, filename=None, md=None, force_compress=False,  nobytes=2,
-                         with_pickle=True ): 
-    
-    ''' YG.Dev@CHX Nov 20, 2017
-        Compress the timepixeldata, in a format of x, y, t
-        x: pos_x in pixel
-        y: pos_y in pixel
-        timepix3 det size 256, 256
-        TODOLIST: mask is not working now
-        Input:
-          pos: 256 * y + x         
-          t: arrival time in sec   
-          filename: the output filename
-          md: a dict to describle the data info   
-          force_compress: if False, 
-                              if already compressed, just it 
-                              else: compress 
-                          if True, compress and, if exist, overwrite the already-coompress data
-        Return:
-          N (frame number)
-            
-    '''     
-    if filename is None:
-        filename=  '/XF11ID/analysis/Compressed_Data' +'/timpix_uid_%s.cmp'%md['uid'] 
-        
-    if force_compress:
-        print ("Create a new compress file with filename as :%s."%filename)
-        return init_compress_timepix_data(   pos, t, tbins, filename=filename, md=md, nobytes= nobytes,
-                                         with_pickle=with_pickle )        
-    else:
-        if not os.path.exists( filename ):
-            print ("Create a new compress file with filename as :%s."%filename)
-            return init_compress_timepix_data(   pos, t, tbins, filename=filename, md=md, nobytes= nobytes,
-                                             with_pickle=with_pickle )
-        else:      
-            print ("Using already created compressed file with filename as :%s."%filename)
-            return    pkl.load( open(filename + '.pkl', 'rb' ) ) 
-            
-            #FD = Multifile(filename, 0, int(1e25)  )       
-            #return    get_FD_end_num(FD)
+from pyCHX.chx_libs import multi_tau_lags
 
 
-            
-            
 
-def create_timepix_compress_header( md, filename, nobytes=2, bins=1  ):
-    '''
-    Create the head for a compressed eiger data, this function is for parallel compress
-    '''    
-    fp = open( filename,'wb' )
-    #Make Header 1024 bytes   
-    #md = images.md
-    if bins!=1:
-        nobytes=8        
-    Header = struct.pack('@16s8d7I916x',b'Version-COMPtpx1',
-                    md['beam_center_x'],md['beam_center_y'], md['count_time'], md['detector_distance'],
-                    md['frame_time'],md['incident_wavelength'], md['x_pixel_size'],md['y_pixel_size'],
-                         
-            nobytes, md['sy'], md['sx'],
-            0,256,
-            0,256
-        )     
-    fp.write( Header)      
-    fp.close()     
-        
-            
-            
-            
-            
-        
-        
-def init_compress_timepix_data( pos, t,  tbins, filename, mask=None,
-                           md = None, nobytes=2,with_pickle=True   ):    
-    ''' YG.Dev@CHX Nov 19, 2017
-        Compress the timepixeldata, in a format of x, y, t
-        x: pos_x in pixel
-        y: pos_y in pixel
-        timepix3 det size 256, 256
-        TODOLIST: mask is not working now
-        Input:
-          pos: 256 * x + y   #can't be 256*x + y        
-          t: arrival time in sec    
-          filename: the output filename
-          md: a dict to describle the data info    
-        Return:
-          N (frame number)
-            
-    ''' 
-    fp = open( filename,'wb' ) 
-    if md is None:
-        md={}
-        md['beam_center_x'] = 0
-        md['beam_center_y'] = 0
-        md['count_time'] = 0
-        md['detector_distance'] = 0
-        md['frame_time'] = 0
-        md['incident_wavelength'] =0
-        md['x_pixel_size'] = 45
-        md['y_pixel_size'] = 45
-        #nobytes = 2
-        md['sx'] = 256
-        md['sy'] = 256    
- 
-     
-    #TODList: for different detector using different md structure, March 2, 2017,
-    
-    #8d include, 
-    #'bytes', 'nrows', 'ncols', (detsize)
-    #'rows_begin', 'rows_end', 'cols_begin', 'cols_end'  (roi)                        
-    Header = struct.pack('@16s8d7I916x',b'Version-COMPtpx1',
-                    md['beam_center_x'],md['beam_center_y'], md['count_time'], md['detector_distance'],
-                    md['frame_time'],md['incident_wavelength'], md['x_pixel_size'],md['y_pixel_size'],
-                         
-            nobytes, md['sy'], md['sx'],
-            0,256,
-            0,256
-        )     
-    fp.write( Header)
-    
-    tx = np.arange(  t.min(), t.max(),  tbins   )      
-    N = len(tx)
-    imgsum  =  np.zeros(   N-1   ) 
-    print('There are %s frames to be compressed...'%N)
-    good_count = 0
-    avg_img = np.zeros(  [ md['sy'], md['sx'] ], dtype= np.float ) 
-    for i in tqdm( range(N-1) ):
-        ind1 = np.argmin( np.abs( tx[i] - t)  )
-        ind2 = np.argmin( np.abs( tx[i+1] - t )  )
-        #print( 'N=%d:'%i, ind1, ind2 )            
-        p_i = pos[ind1: ind2]
-        ps,vs = np.unique( p_i, return_counts= True )  
-        np.ravel(avg_img )[ps] +=   vs
-        good_count +=1             
-        dlen = len(ps)   
-        imgsum[i] = vs.sum()    
-        #print(vs.sum())
-        fp.write(  struct.pack( '@I', dlen   ))
-        fp.write(  struct.pack( '@{}i'.format( dlen), *ps))
-        fp.write(  struct.pack( '@{}{}'.format( dlen,'ih'[nobytes==2]), *vs)) 
-    fp.close()
-    avg_img /= good_count
-    #return N -1
-    if  with_pickle:
-        pkl.dump( [  avg_img, imgsum, N-1 ], open(filename + '.pkl', 'wb' ) )      
-    return   avg_img, imgsum, N-1 
-    
-    
-    
-    
-    
-
-def compress_timepix_data_old( data_pixel, filename, rois=None,
-                           md = None, nobytes=2   ):    
+def compress_timepixeldata( p,t, tbins, filename,  md= None,nobytes=2   ):    
     '''
         Compress the timepixeldata
-        md: a dict to describle the data info
-        rois: [y1,y2, x1, x2]
+        md: a string to describle the data info
+ 
             
     ''' 
-    fp = open( filename,'wb' ) 
+    fp = open( filename,'wb' )
+ 
     if md is None:
         md={}
         md['beam_center_x'] = 0
@@ -202,35 +40,41 @@ def compress_timepix_data_old( data_pixel, filename, rois=None,
         md['y_pixel_size'] =25
         #nobytes = 2
         md['sx'] = 256
-        md['sy'] = 256  
-        md['roi_rb']= 0
-        md['roi_re']= md['sy']
-        md['roi_cb']= 0
-        md['roi_ce']= md['sx']
-    if rois is not None:
-        md['roi_rb']= rois[2]
-        md['roi_re']= rois[3]
-        md['roi_cb']= rois[1]
-        md['roi_ce']= rois[0]
-        
-        md['sy'] = md['roi_cb'] - md['roi_ce']
-        md['sx'] = md['roi_re'] - md['roi_rb']
-     
-    #TODList: for different detector using different md structure, March 2, 2017,
-    
-    #8d include, 
-    #'bytes', 'nrows', 'ncols', (detsize)
-    #'rows_begin', 'rows_end', 'cols_begin', 'cols_end'  (roi)                        
-    Header = struct.pack('@16s8d7I916x',b'Version-COMPtpx1',
+        md['sy'] = 256    
+                            
+    Header = struct.pack('@16s8d7I916x',b'Version-COMP0001',
                     md['beam_center_x'],md['beam_center_y'], md['count_time'], md['detector_distance'],
                     md['frame_time'],md['incident_wavelength'], md['x_pixel_size'],md['y_pixel_size'],
-                         
-            nobytes, md['sy'], md['sx'],
-            md['roi_rb'], md['roi_re'],md['roi_cb'],md['roi_ce']
-        )
+                        nobytes, md['sy'], md['sx'],
+                         0, md['sy'],
+                         0, md['sx']                
+                    )
+        
+        
+    #Header = struct.pack('@16s%ds%dx'%(n,lx),b'Version-COMP0002', md.encode('ascii')  )      
+    fp.write( Header)
      
-    fp.write( Header)  
-    fp.write( data_pixel )
+    tx = np.arange(  t.min(), t.max(), tbins   )  
+    N = len(tx)
+    print( 'Will Compress %s Frames.'%(N-1)  )
+    #imgsum  =  np.zeros(    N   )   
+    for i in tqdm( range(N-1) ):
+        ind1 = np.argmin( np.abs( tx[i] - t )  )
+        ind2 = np.argmin( np.abs( tx[i+1] - t )  )
+        #print( i, ind1, ind2 )
+        p_i = p[ind1: ind2]
+        pos,val = np.unique( p_i, return_counts= True )
+        #print(i,  pos, val )            
+        dlen = len(pos)         
+        #imgsum[i] = val.sum()
+        fp.write(  struct.pack( '@I', dlen   ))
+        fp.write(  struct.pack( '@{}i'.format( dlen), *pos))
+        fp.write(  struct.pack( '@{}{}'.format( dlen,'ih'[nobytes==2]), *val)) 
+    fp.close()  
+    return N -1 
+        #return    imgsum
+
+    
     
     
     
@@ -244,18 +88,16 @@ class Get_TimePixel_Arrayc(object):
         data_pixel =   Get_Pixel_Array( imgsr, pixelist).get_data()
     '''
     
-    def __init__(self, pos, hitime, tbins, pixelist, beg=None, end=None, norm=None,flat_correction=None,
-                 detx = 256, dety = 256):
+    def __init__(self, pos, hitime, tbins, pixelist, beg=None, end=None, norm=None, detx = 256, dety = 256):
         '''
         indexable: a images sequences
         pixelist:  1-D array, interest pixel list
-        #flat_correction, normalized by flatfield
-        #norm, normalized by total intensity, like a incident beam intensity
         '''
         self.hitime = hitime
         self.tbins   = tbins
-        self.tx = np.arange(  self.hitime.min(), self.hitime.max(), self.tbins   )        
-        N = len(self.tx)        
+        self.tx = np.arange(  self.hitime.min(), self.hitime.max(), self.tbins   )  
+        N = len(self.tx)
+        
         if beg is None:
             beg = 0
         if end is None:
@@ -266,10 +108,11 @@ class Get_TimePixel_Arrayc(object):
         self.length = self.end - self.beg 
         self.pos = pos
         self.pixelist = pixelist        
-        self.norm = norm  
-        self.flat_correction = flat_correction
+        self.norm = norm   
         self.detx = detx
         self.dety = dety
+
+        
         
     def get_data(self ): 
         '''
@@ -277,7 +120,7 @@ class Get_TimePixel_Arrayc(object):
         Return: 2-D array, shape as (len(images), len(pixellist))
         '''
         norm = self.norm
-        data_array = np.zeros([ self.length-1,len(self.pixelist)]) 
+        data_array = np.zeros([ self.length,len(self.pixelist)]) 
         print( data_array.shape)
         
         #fra_pix = np.zeros_like( pixelist, dtype=np.float64)
@@ -286,54 +129,59 @@ class Get_TimePixel_Arrayc(object):
         n=0
         tx = self.tx
         N = len(self.tx)
-        print( 'The Produced Array Length is  %d.'%(N-1)  )
-        flat_correction = self.flat_correction
+        print( 'Will Compress %s Frames.'%(N-1)  )
         #imgsum  =  np.zeros(    N   )   
         for i in tqdm( range(N-1) ):
             ind1 = np.argmin( np.abs( tx[i] - self.hitime )  )
             ind2 = np.argmin( np.abs( tx[i+1] - self.hitime )  )
-            #print( 'N=%d:'%i, ind1, ind2 )            
+            
+            #print( i, ind1, ind2 )
+            
             p_i = self.pos[ind1: ind2]
-            pos,val = np.unique( p_i, return_counts= True )            
+            pos,val = np.unique( p_i, return_counts= True )
+            
             #print( val.sum() )
+
             w = np.where( timg[pos] )[0]
             pxlist = timg[  pos[w]   ] -1 
             #print( val[w].sum() )
-            #fra_pix[ pxlist] = v[w]                
-            if flat_correction is not None:
-                #normalized by flatfield
+            #fra_pix[ pxlist] = v[w] 
+            if norm is None:
                 data_array[n][ pxlist] = val[w] 
-            else:
-                data_array[n][ pxlist] = val[w] / flat_correction[pxlist]   #-1.0
-            if norm is not None: 
-                #normalized by total intensity, like a incident beam intensity
-                data_array[n][ pxlist] /= norm[i] 
-            n += 1            
+            else: 
+                data_array[n][ pxlist] = val[w] / norm[pxlist]   #-1.0                    
+            n += 1
+            
         return data_array     
     
-   
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
     
 def apply_timepix_mask( x,y,t, roi ):
     y1,y2, x1,x2  = roi
     w =  (x < x2) & (x >= x1) & (y < y2) & (y >= y1) 
     return x[w],y[w], t[w]
 
-
-
-def get_timepixel_data( data_dir, filename, time_unit=6.1e-12):
+def get_timepixel_data( data_dir, filename):
     '''give a csv file of a timepixel data, return x,y,t (in ps/6.1)
-    x, pos_x in pixel
-    y, pos_y in pixel
-    t, arrival time
-    time_unit, t*time_unit will convert to second, by default is 6.1e-12
-    
-    return x,y,t (in second, starting from zero)       
+        y1,y2, x1,x2 = roi   
     '''
     data =  pds.read_csv( data_dir + filename )
-    #'#Col', ' #Row', ' #ToA',
-    #return np.array( data['Col'] ), np.array(data['Row']), np.array(data['GlobalTimeFine']) #*6.1  #in ps
-    x,y,t=np.array( data['#Col'] ), np.array(data[' #Row']), np.array(data[' #ToA'] )  * time_unit
-    return x,y,  t-t.min() #* 25/4096.  #in ns   
+    return data['Col'], data['Row'], data['GlobalTimeFine'] #*6.1  #in ps   
 
 
 def get_timepixel_data_from_series( data_dir, filename_prefix, 
@@ -358,24 +206,17 @@ def get_timepixel_data_from_series( data_dir, filename_prefix,
         
      
 
-def get_timepixel_avg_image( x,y,t,  det_shape = [256, 256], delta_time = None   ):
-    '''YG.Dev@CHX, 2016
-    give x,y, t data to get image in a period of delta_time (in second)
-    x, pos_x in pixel
-    y, pos_y in pixel
-    t, arrival time
-    
-    
-    '''
-    t0 = t.min() 
-    tm = t.max() 
+def get_timepixel_image( x,y,t, det_shape = [256, 256], delta_time = None   ):
+    '''give x,y, t data to get image in a period of delta_time (in second)'''
+    t0 = t.min() *6.1
+    tm = t.max() *6.1
     
     if delta_time is not None:
         delta_time *=1e12
         if delta_time > tm:
             delta_time = tm            
     else:
-        delta_time = t.max()
+        delta_time = tm
     #print( delta_time)
     t_ = t[t<delta_time]
     x_ = x[:len(t_)]
@@ -388,7 +229,7 @@ def get_timepixel_avg_image( x,y,t,  det_shape = [256, 256], delta_time = None  
     print( 'The max photon count is %d.'%img.max())
     return img
     
-def get_his_taus( t,  bin_step  ):
+def get_his_taus( t,  bin_step, time_unit= 1.0 ):
     '''Get taus and  histrogram of photons
        Parameters:
             t: the time stamp of photon hitting the detector
@@ -399,9 +240,9 @@ def get_his_taus( t,  bin_step  ):
     
     '''
     
-    bins = np.arange(  t.min(), t.max(), bin_step   )  #1e6 for us
+    bins = np.arange(  t.min()*time_unit , t.max()*time_unit, bin_step   )  #1e6 for us
     #print( bins )
-    td = np.histogram( t, bins=bins )[0] #do histogram
+    td = np.histogram( t*time_unit, bins=bins )[0] #do histogram
     taus = (bins - bins[0])  
     return taus[1:], td
 
@@ -644,20 +485,6 @@ class xpcs( object):
         plt.show()
     
  
-def histogram_xyt( x, y, t, binstep=100, detx=256, dety=256  ):
-    '''x: coordinate-x
-       y: coordinate-y
-       t: photon hit time
-       
-       bin t in binstep (in t unit) period
-    
-    '''    
-    L= np.max( (t-t[0])//binstep ) + 1
-    arr = np.ravel_multi_index( [x, y, (t-t[0])//binstep ], [detx, dety,L ]    )
-    M,N = arr.max(),arr.min()
-    da = np.zeros( [detx, dety, L ]  )
-    da.flat[np.arange(N,  M  ) ] = np.bincount( arr- N  )
-    return da
 
 ######################################################
