+++ requirements.txt
@@ -1,15 +1,15 @@
-numpy
 cython
-matplotlib
-scipy
-lmfit
-pillow
-tifffile
-pyyaml
-pandas
 databroker
+lmfit
+matplotlib
+numpy
+pandas
+pillow
+pyyaml
+scipy
+tifffile
 
 git+https://github.com/scikit-beam/scikit-beam.git#egg=scikit-beam
-git+https://github.com/nsls-ii-chx/eiger-io.git#eiger-io
+git+https://github.com/NSLS-II-CHX/eiger-io.git#eiger-io
 git+https://github.com/NSLS-II-CHX/chxtools.git#egg=chxtools
 git+https://github.com/Nikea/xray-vision.git#Xray-vision
