+++ setup.cfg
@@ -1,6 +1,6 @@
 [versioneer]
 VCS=git
 style=pep440
-versionfile_source=chxanalys/_version.py
-versionfile_build=chxanalys/_version.py
+versionfile_source=pyCHX/_version.py
+versionfile_build=pyCHX/_version.py
 tag_prefix=v
